test setup file for add services updated

+++ payments/src/test/setup.ts
@@ -22,8 +22,8 @@ beforeAll(async () => {
 
 beforeEach(async () => {
   jest.clearAllMocks();
-  const collections = await mongoose.connection.db!.collections();
-
+  const collections = await mongoose.connection.db?.collections();
+  if (!collections) return
   for (let collection of collections) {
     await collection.deleteMany({});
   }
